scsi: qedf: Use designated initializer for struct qed_fcoe_cb_ops
Recent fixes to the randstruct GCC plugin allowed it to notice that this
structure is entirely function pointers and is therefore subject to
randomization, but doing so requires that it always use designated
initializers. Explicitly specify the "common" member as being
initialized. Silences:

drivers/scsi/qedf/qedf_main.c:702:9: error: positional initialization of field in 'struct' declared with 'designated_init' attribute [-Werror=designated-init]
  702 |         {
      |         ^

Fixes: c2ea09b193d2 ("randstruct: gcc-plugin: Remove bogus void member")
Signed-off-by: Kees Cook <kees@kernel.org>
Link: https://lore.kernel.org/r/20250502224156.work.617-kees@kernel.org
Signed-off-by: Martin K. Petersen <martin.petersen@oracle.com>